Fixed compilation errors and other issues

+++ wrapper/cncvis.pxd
@@ -7,5 +7,5 @@ cdef extern from "../src/ncvis.hpp" namespace "ncvis":
 
 cdef extern from "../src/ncvis.hpp" namespace "ncvis":
     cdef cppclass NCVis:
-        NCVis(long d, long n_threads, long n_neighbors, long M, long ef_construction, long random_seed, int n_epochs, int n_init_epochs, float a, float b, float alpha, float alpha_Q, long* n_noise, Distance dist) except +
-        void fit_transform(const float *const X, long N, long D, float* Y)
+        NCVis(size_t d, size_t n_threads, size_t n_neighbors, size_t M, size_t ef_construction, size_t random_seed, int n_epochs, int n_init_epochs, float a, float b, float alpha, float alpha_Q, size_t* n_noise, Distance dist) except +
+        void fit_transform(const float *const X, size_t N, size_t D, float* Y)
